+++ clients/fides-js/src/fides.ts
@@ -47,57 +47,20 @@
 import { gtm } from "./integrations/gtm";
 import { meta } from "./integrations/meta";
 import { shopify } from "./integrations/shopify";
-import { getConsentContext } from "./lib/consent-context";
-import { initOverlay } from "./lib/consent";
-import {
-  CookieKeyConsent,
-  CookieIdentity,
-  CookieMeta,
-  getOrMakeFidesCookie,
-  makeConsentDefaultsLegacy,
-  buildCookieConsentForExperiences,
-  FidesCookie,
-  isNewFidesCookie,
-  updateExperienceFromCookieConsent,
-} from "./lib/cookie";
-import {
-  PrivacyExperience,
-  FidesConfig,
-  FidesOptions,
-  UserGeolocation,
-  ConsentMethod,
-  SaveConsentPreference,
-  ConsentMechanism,
-  EmptyExperience,
-} from "./lib/consent-types";
-import {
-  constructFidesRegionString,
-  debugLog,
-  experienceIsValid,
-  transformConsentToFidesUserPreference,
-  validateOptions,
-  isPrivacyExperience,
-} from "./lib/consent-utils";
-import { dispatchFidesEvent } from "./lib/events";
-import { fetchExperience } from "./services/fides/api";
-import { getGeolocation } from "./services/external/geolocation";
-import { OverlayProps } from "./components/Overlay";
-import { updateConsentPreferences } from "./lib/preferences";
-import { resolveConsentValue } from "./lib/consent-value";
 
-export type Fides = {
-  consent: CookieKeyConsent;
-  experience?: PrivacyExperience | EmptyExperience;
-  geolocation?: UserGeolocation;
-  options: FidesOptions;
-  fides_meta: CookieMeta;
-  gtm: typeof gtm;
-  identity: CookieIdentity;
-  init: typeof init;
-  initialized: boolean;
-  meta: typeof meta;
-  shopify: typeof shopify;
-};
+import { isNewFidesCookie } from "./lib/cookie";
+import { FidesConfig } from "./lib/consent-types";
+
+import { dispatchFidesEvent } from "./lib/events";
+
+import {
+  initialize,
+  getInitialCookie,
+  getInitialFides,
+} from "./lib/initialize";
+import type { Fides } from "./lib/initialize";
+
+import { renderOverlay } from "./lib/renderOverlay";
 
 declare global {
   interface Window {
@@ -109,203 +72,28 @@ declare global {
 // eslint-disable-next-line no-underscore-dangle,@typescript-eslint/naming-convention
 let _Fides: Fides;
 
-const retrieveEffectiveRegionString = async (
-  geolocation: UserGeolocation | undefined,
-  options: FidesOptions
-) => {
-  // Prefer the provided geolocation if available and valid; otherwise, fallback to automatically
-  // geolocating the user by calling the geolocation API
-  const fidesRegionString = constructFidesRegionString(geolocation);
-  if (!fidesRegionString) {
-    // we always need a region str so that we can PATCH privacy preferences to Fides Api
-    return constructFidesRegionString(
-      // Call the geolocation API
-      await getGeolocation(
-        options.isGeolocationEnabled,
-        options.geolocationApiUrl,
-        options.debug
-      )
-    );
-  }
-  return fidesRegionString;
-};
-
-const automaticallyApplyGPCPreferences = (
-  cookie: FidesCookie,
-  fidesRegionString: string | null,
-  fidesApiUrl: string,
-  debug: boolean,
-  effectiveExperience?: PrivacyExperience
-) => {
-  if (!effectiveExperience || !effectiveExperience.privacy_notices) {
-    return;
-  }
-
-  const context = getConsentContext();
-  if (!context.globalPrivacyControl) {
-    return;
-  }
-
-  let gpcApplied = false;
-  const consentPreferencesToSave = effectiveExperience.privacy_notices.map(
-    (notice) => {
-      if (
-        notice.has_gpc_flag &&
-        !notice.current_preference &&
-        notice.consent_mechanism !== ConsentMechanism.NOTICE_ONLY
-      ) {
-        gpcApplied = true;
-        return new SaveConsentPreference(
-          notice,
-          transformConsentToFidesUserPreference(false, notice.consent_mechanism)
-        );
-      }
-      return new SaveConsentPreference(
-        notice,
-        transformConsentToFidesUserPreference(
-          resolveConsentValue(notice, context),
-          notice.consent_mechanism
-        )
-      );
-    }
-  );
-
-  if (gpcApplied) {
-    updateConsentPreferences({
-      consentPreferencesToSave,
-      experienceId: effectiveExperience.id,
-      fidesApiUrl,
-      consentMethod: ConsentMethod.gpc,
-      userLocationString: fidesRegionString || undefined,
-      cookie,
-      debug,
-    });
-  }
-};
-
 /**
  * Initialize the global Fides object with the given configuration values
  */
-const init = async ({
-  consent,
-  experience,
-  geolocation,
-  options,
-}: FidesConfig) => {
-  // Configure the default legacy consent values
-  const context = getConsentContext();
-  const consentDefaults = makeConsentDefaultsLegacy(
-    consent,
-    context,
-    options.debug
-  );
-
-  // Load any existing user preferences from the browser cookie
-  const cookie: FidesCookie = getOrMakeFidesCookie(
-    consentDefaults,
-    options.debug
-  );
-
-  // If saved preferences are detected, immediately initialize from local cache,
-  // and then continue geolocating, etc.
-  const hasExistingCookie = !isNewFidesCookie(cookie);
-
-  if (hasExistingCookie) {
-    _Fides.consent = cookie.consent;
-    _Fides.fides_meta = cookie.fides_meta;
-    _Fides.identity = cookie.identity;
-    _Fides.experience = experience;
-    _Fides.geolocation = geolocation;
-    _Fides.options = options;
-    _Fides.initialized = true;
-    if (isPrivacyExperience(experience)) {
-      // at this point, pre-fetched experience contains no user consent, so we populate with the Fides cookie
-      updateExperienceFromCookieConsent(experience, cookie, options.debug);
-    }
-    dispatchFidesEvent("FidesInitialized", cookie, options.debug);
-    dispatchFidesEvent("FidesUpdated", cookie, options.debug);
+const init = async (config: FidesConfig) => {
+  const cookie = getInitialCookie(config);
+  const initialFides = getInitialFides({ ...config, cookie });
+  if (initialFides) {
+    Object.assign(_Fides, initialFides);
+    dispatchFidesEvent("FidesInitialized", cookie, config.options.debug);
+    dispatchFidesEvent("FidesUpdated", cookie, config.options.debug);
   }
-
-  let shouldInitOverlay: boolean = options.isOverlayEnabled;
-  let effectiveExperience = experience;
-  let fidesRegionString: string | null = null;
-
-  if (shouldInitOverlay) {
-    if (!validateOptions(options)) {
-      debugLog(
-        options.debug,
-        "Invalid overlay options. Skipping overlay initialization.",
-        options
-      );
-      shouldInitOverlay = false;
-    }
-
-    fidesRegionString = await retrieveEffectiveRegionString(
-      geolocation,
-      options
-    );
-
-    if (!fidesRegionString) {
-      debugLog(
-        options.debug,
-        `User location could not be obtained. Skipping overlay initialization.`
-      );
-      shouldInitOverlay = false;
-      // if the experience object is null or empty from the server, we should fetch client side
-    } else if (!isPrivacyExperience(experience)) {
-      effectiveExperience = await fetchExperience(
-        fidesRegionString,
-        options.fidesApiUrl,
-        options.debug,
-        cookie.identity.fides_user_device_id
-      );
-    }
-
-    if (experienceIsValid(effectiveExperience, options)) {
-      // Overwrite cookie consent with experience-based consent values
-      cookie.consent = buildCookieConsentForExperiences(
-        effectiveExperience as PrivacyExperience,
-        context,
-        options.debug
-      );
-
-      if (shouldInitOverlay) {
-        await initOverlay(<OverlayProps>{
-          experience: effectiveExperience,
-          fidesRegionString,
-          cookie,
-          options,
-        }).catch(() => {});
-      }
-    }
-  }
-  if (shouldInitOverlay && isPrivacyExperience(effectiveExperience)) {
-    automaticallyApplyGPCPreferences(
-      cookie,
-      fidesRegionString,
-      options.fidesApiUrl,
-      options.debug,
-      effectiveExperience
-    );
-  }
-
-  // Initialize the window.Fides object
-  _Fides.consent = cookie.consent;
-  _Fides.fides_meta = cookie.fides_meta;
-  _Fides.identity = cookie.identity;
-  _Fides.experience = experience;
-  _Fides.geolocation = geolocation;
-  _Fides.options = options;
-  _Fides.initialized = true;
+  const updatedFides = await initialize({ ...config, cookie, renderOverlay });
+  Object.assign(_Fides, updatedFides);
 
   // Dispatch the "FidesInitialized" event to update listeners with the initial
   // state. Skip if we already initialized due to an existing cookie.
   // For convenience, also dispatch the "FidesUpdated" event; this allows
   // listeners to ignore the initialization event if they prefer
-  if (!hasExistingCookie) {
-    dispatchFidesEvent("FidesInitialized", cookie, options.debug);
+  if (isNewFidesCookie(cookie)) {
+    dispatchFidesEvent("FidesInitialized", cookie, config.options.debug);
   }
-  dispatchFidesEvent("FidesUpdated", cookie, options.debug);
+  dispatchFidesEvent("FidesUpdated", cookie, config.options.debug);
 };
 
 // The global Fides object; this is bound to window.Fides if available
@@ -350,6 +138,3 @@ export * from "./lib/consent-utils";
 export * from "./lib/consent-value";
 export * from "./lib/cookie";
 export * from "./lib/events";
-
-// DEFER: this default export isn't very useful, it's just the Fides type
-export default Fides;